+++ htdocs/js/init.js
@@ -246,6 +246,7 @@ function appInit() {
     //hide popover
     domCache.body.addEventListener('click', function() {
         hidePopover();
+        hideDropdown(event);
     }, false);
     //init modules
     initGlobalModals();