+++ src/lib/list.c
@@ -101,6 +101,25 @@ void list_free_cb_sds_user_data(struct t_list_node *current) {
 }
 
 /**
+ * Gets a list node index by key.
+ * @param l list
+ * @param key key to get
+ * @return int index of the key, -1 if not found
+ */
+int list_get_node_idx(const struct t_list *l, const char *key) {
+    struct t_list_node *current = l->head;
+    int i = 0;
+    while (current != NULL) {
+        if (strcmp(current->key, key) == 0) {
+            return i;
+        }
+        current = current->next;
+        i++;
+    }
+    return -1;
+}
+
+/**
  * Gets a list node by key.
  * @param l list
  * @param key key to get