+++ src/lib/mympd_state.h
@@ -162,6 +162,7 @@ struct t_partition_state {
     sds stream_uri;                        //!< custom url for local playback
     //lists
     struct t_list last_played;             //!< last_played list
+    struct t_list presets;                 //!< Playback presets
 };
 
 /**